requirements.txt
Flask
gevent
gevent-websocket
Flask-SocketIO
socketIO-client-2